+++ CHANGELOG.md
@@ -10,6 +10,8 @@ and this project adheres to [Semantic Versioning](http://semver.org/).
 - fix memory leak when exporting a big amount of components
 - fix running import command from a non-root directory
 - support specifying multiple ids using export command 
+- fix the auto creating dependencies during commit
+- performance improvement for status and commit
 
 ## [0.10.2] - 2017-08-07
 Improve resolving packages dependencies for ts files