+++ CHANGELOG.md
@@ -7,6 +7,8 @@ and this project adheres to [Semantic Versioning](http://semver.org/).
 
 ## [unreleased]
 
+## [0.10.3] - 2017-08-08
+
 - fix memory leak when exporting a big amount of components
 - fix running import command from a non-root directory
 - support specifying multiple ids using export command